overflows and stuff

+++ gateComs/LEDUtils.py
@@ -299,7 +299,7 @@ class LEDStrip:
         c = [2,0.25,0]
         gain = c[0]+c[1]+c[2]
         for j in range(self.strip.numPixels()):
-            v = round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75)
+            v = int(round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75))
             c = [c[0]*gain*v,c[1]*gain*v,c[2]*gain*v]
             self.strip.setPixelColor(j, Color(int(c[0]),int(c[1]),int(c[2])))
         self.strip.show()
@@ -315,7 +315,7 @@ class LEDStrip:
         c = [1,0.5,1]
         gain = c[0]+c[1]+c[2]
         for j in range(self.strip.numPixels()):
-            v = round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75)
+            v = int(round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75))
             c = [c[0]*gain*v,c[1]*gain*v,c[2]*gain*v]
             self.strip.setPixelColor(j, Color(int(c[0]),int(c[1]),int(c[2])))
         self.strip.show()
@@ -331,7 +331,7 @@ class LEDStrip:
         c = [0.5,0,1]
         gain = c[0]+c[1]+c[2]
         for j in range(self.strip.numPixels()):
-            v = round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75)
+            v = int(round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75))
             c = [c[0]*gain*v,c[1]*gain*v,c[2]*gain*v]
             self.strip.setPixelColor(j, Color(int(c[0]),int(c[1]),int(c[2])))
         self.strip.show()
@@ -347,7 +347,7 @@ class LEDStrip:
         c = [0,1,0.75]
         gain = c[0]+c[1]+c[2]
         for j in range(self.strip.numPixels()):
-            v = round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75)
+            v = int(round((math.sin((x*math.pi*(2.0/frameCount))-(math.pi*.5))+1)*127.5*.75))
             c = [c[0]*gain*v,c[1]*gain*v,c[2]*gain*v]
             self.strip.setPixelColor(j, Color(int(c[0]),int(c[1]),int(c[2])))
         self.strip.show()
